requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
MarkupSafe==2.1.2
orjson==3.9.0
packaging==23.1
passlib==1.7.4
pip==23.1.2
pluggy==1.0.0
pyasn1==0.5.0
pydantic==1.10.7
pytest==7.3.1
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.26.1
typing_extensions==4.5.0
ujson==5.7.0
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.2
alembic==1.11.1
bcrypt==4.0.1